+++ home/models/__init__.py
@@ -1,0 +1,4 @@
+from home.models.contact import Contact, ContactKind
+from home.models.main import GroupInfo, User, UserToken
+from home.models.notice import Notice
+from home.models.notification import Notification, NotificationRead
