+++ home/models/main.py
@@ -224,208 +224,3 @@ class GroupInfo(models.Model):
     order = models.PositiveIntegerField(
         default=0
     )
-
-
-class Notice(models.Model):
-    """ ホーム画面に表示される「お知らせ」
-    """
-
-    # settings
-
-    class Meta:
-        verbose_name = 'お知らせ'
-        verbose_name_plural = verbose_name
-
-    def __str__(self):
-        return self.title
-
-    # fields
-
-    title = models.CharField(
-        verbose_name='タイトル',
-        max_length=30,
-        help_text='30 文字以内で入力してください。'
-    )
-
-    body = models.TextField(
-        verbose_name='本文',
-        max_length=100,
-        help_text='100 文字以内で入力してください。あまり長い文章は好ましくありません。'
-    )
-
-    writer = models.ForeignKey(
-        'home.User',
-        verbose_name='入力者',
-        on_delete=models.SET_NULL,
-        null=True
-    )
-
-    update_datetime = models.DateTimeField(
-        verbose_name='最終更新日時',
-        auto_now=True
-    )
-
-
-class Contact(models.Model):
-    """お問い合わせ
-    """
-
-    # settings
-
-    class Meta:
-        verbose_name = 'お問い合わせ'
-        verbose_name_plural = verbose_name
-
-    def __str__(self):
-        return self.title
-
-    # fields
-    title = models.CharField(
-        verbose_name='表題',
-        max_length=100
-    )
-
-    body = models.TextField(
-        verbose_name='本文',
-        max_length=2000
-    )
-
-    kind = models.ForeignKey(
-        'home.ContactKind',
-        verbose_name='お問い合わせ種別',
-        on_delete=models.CASCADE
-    )
-
-    writer = models.ForeignKey(
-        'home.User',
-        verbose_name='入力者',
-        on_delete=models.CASCADE
-    )
-
-    create_datetime = models.DateTimeField(
-        verbose_name='作成日時',
-        auto_now_add=True
-    )
-
-
-class ContactKind(models.Model):
-    """お問い合わせ種別
-    """
-
-    # settings
-
-    class Meta:
-        verbose_name = 'お問い合わせ種別'
-        verbose_name_plural = verbose_name
-        ordering = ('order',)
-
-    def __str__(self):
-        return self.name
-
-    # fields
-
-    name = models.CharField(
-        verbose_name='種別名',
-        max_length=100
-    )
-
-    groups = models.ManyToManyField(
-        'auth.Group',
-        verbose_name='管轄'
-    )
-
-    order = models.PositiveIntegerField(
-        default=0
-    )
-
-
-class Notification(models.Model):
-    """通知
-    """
-
-    # settings
-
-    class Meta:
-        verbose_name = '通知'
-        verbose_name_plural = verbose_name
-
-    def __str__(self):
-        return self.title
-
-    def has_read(self, user):
-        """ある User がすでに通知を呼んだかどうかを判定する
-
-        NotificationRead がヒットするかどうかを返す
-        """
-        return self.notificationread_set.filter(
-            user=user
-        ).exists()
-
-    # fields
-
-    title = models.CharField(
-        verbose_name='表題',
-        max_length=50
-    )
-
-    body = models.TextField(
-        verbose_name='本文',
-        max_length=2000
-    )
-
-    to = models.ManyToManyField(
-        'home.User',
-        verbose_name='宛先',
-        related_name='notification_to'
-    )
-
-    sender = models.ForeignKey(
-        'home.User',
-        verbose_name='送信者',
-        on_delete=models.CASCADE,
-        related_name='notification_from'
-    )
-
-    group = models.ForeignKey(
-        'auth.Group',
-        verbose_name='担当',
-        on_delete=models.CASCADE,
-        help_text='どの担当として送信するかをここで選択してください。',
-    )
-
-    create_datetime = models.DateTimeField(
-        verbose_name='送信日時',
-        auto_now_add=True
-    )
-
-
-class NotificationRead(models.Model):
-    """通知既読情報
-    """
-
-    # settings
-
-    class Meta:
-        verbose_name = '通知既読'
-        verbose_name_plural = verbose_name
-
-    def __str__(self):
-        return str(self.notification)
-
-    # fields
-    notification = models.ForeignKey(
-        'home.Notification',
-        verbose_name='通知',
-        on_delete=models.CASCADE
-    )
-
-    user = models.ForeignKey(
-        'home.User',
-        verbose_name='ユーザー',
-        on_delete=models.CASCADE
-    )
-
-    create_datetime = models.DateTimeField(
-        verbose_name='開封日時',
-        auto_now_add=True
-    )
